+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Reverted config for HTTP and HTTPS ports, cause it is not needed.
 * Auto cert http server will now error out if it fails.
 * Enabled tls-alpn ACME challenges for TLS.
+* Switched from TOML to INI config format because it's more standard.
 
 ## v1.0.0-rc5
 * Hopefully fixed autocert not working on non default ports.